requirements.txt
contourpy==1.2.0
cycler==0.11.0
fonttools==4.25.0
glfw==2.7.0
kiwisolver==1.4.4
matplotlib==3.8.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
munkres==1.1.4
numpy==1.26.4
packaging==23.2
pillow~=10.3.0
pip==23.3.1
ply==3.11
PyGLM==2.7.1
PyOpenGL==3.1.7
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-dateutil==2.8.2
setuptools==68.2.2
sip==6.7.12
six==1.16.0
tomli==2.0.1
tornado==6.3.3
wheel==0.41.2

websockets~=12.0
websocket-client~=1.7.0